feat: pca switch
