+++ Cargo.toml
@@ -47,6 +47,7 @@ import = ["bevy_args", "burn-import", "clap", "serde"]
 bevy_args = { git = "https://github.com/mosure/bevy_args.git", branch = "burn", optional = true }
 burn-import = { version = "0.15", features = ["pytorch"], optional = true }
 clap = { version = "4.5", features = ["derive"], optional = true }
+ndarray = "0.16"
 serde = { version = "1.0", optional = true }
 
 
@@ -110,7 +111,6 @@ bhtsne = "0.5.3"
 criterion = { version = "0.5", features = ["html_reports"] }
 futures-intrusive = { version = "0.5.0" }
 image = { version = "0.25", default-features = false, features = ["png"] }
-ndarray = "0.16"
 pollster = { version = "0.4.0" }
 safetensors = "0.4"
